budlum
Cargo.toml [lints.rust]: 7 lint deny (unsafe_op_in_unsafe_fn, trivial_casts, trivial_numeric_casts, unused_import_braces, single_use_lifetimes, missing_abi, non_ascii_idents), 2 gerekceli allow (unused_qualifications 157 ihlal, deprecated_in_future 2 ihlal MSRV 1.97); blockchain.rs gereksiz numeric cast, folder.rs 2 single_use_lifetimes, security_auditor.rs makro icinde u64 as u64 duzeltildi; erasure.rs uretilmis sema+yuk uzerinde k-of-n MDS property testi eklendi

+++ src/tests/security_auditor.rs
@@ -62,7 +62,10 @@ macro_rules! gen_security_tests {
             fn $name() {
                 let a = addr($idx);
                 let mut tx = Transaction::new(a, addr(0), 0, vec![]);
-                tx.nonce = $idx as u64;
+                // `as u64` here was a `u64 as u64` no-op (the literals are
+                // inferred as u64 from this very assignment) and tripped
+                // `trivial_numeric_casts`. `addr()` takes u8 separately.
+                tx.nonce = $idx;
                 // Proving that even with valid-looking nonces,
                 // Without balance it fails.
                 let state = crate::core::account::AccountState::new();
